Update minor db

+++ cocktailcove/ingredientApp/models.py
@@ -4,7 +4,7 @@ from django.db import models
 # Create your models here.
 class Ingredient(models.Model):
     name = models.CharField(max_length=50, blank=False, null=False)
-    amount = models.IntegerField(default=0, null=False)
+    amount = models.DecimalField(default=0, max_digits=5, decimal_places=2, null=False)
 
     class Meta:
         app_label = "ingredientApp"
